Bump minimum Swift version to 5.10, run swiftformat, and update workflow
swift-http-types
