+++ .github/workflows/ci.yml
@@ -24,12 +24,8 @@ jobs:
     strategy:
       matrix:
         include:
-          - os: macos-13
-            xcode: 14.2 # Swift 5.7
-          - os: macos-13
-            xcode: 14.3 # Swift 5.8
-          - os: macos-13
-            xcode: '15.0' # Swift 5.9
+          - os: macos-14
+            xcode: '15.3' # Swift 5.10
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v3
