+++ .swiftformat
@@ -1,6 +1,6 @@
 --extensionacl on-declarations
 --redundanttype explicit
---swiftversion 5.5
+--swiftversion 5.10
 --maxwidth 120
 --header "{file}\nNetworkService\n\nCopyright © {year} MFB Technologies, Inc. All rights reserved.\n\nThis source code is licensed under the MIT license found in the\nLICENSE file in the root directory of this source tree."
 --allman false
